fixed libraries pathing and freeing depending on OS
